latency_experiment/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
databases[postgresql]>=0.8.0
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=5.0.1
pydantic>=2.5.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.22.0
locust>=2.17.0
numpy>=1.26.0
aiofiles>=23.2.1
psycopg2-binary>=2.9.9
